+++ client/app/src/main/java/edu/cmu/eps/scams/recordings/PhoneCallResult.java
@@ -8,10 +8,10 @@ import java.io.File;
 
 public class PhoneCallResult {
     public final long ringTimestamp;
-    public final int audioLength;
+    public final long audioLength;
     public final File audioRecording;
 
-    public PhoneCallResult(long ringTimestamp, File audioRecording, int audioLength) {
+    public PhoneCallResult(long ringTimestamp, File audioRecording, long audioLength) {
         this.ringTimestamp = ringTimestamp;
         this.audioRecording = audioRecording;
         this.audioLength = audioLength;
